+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Fixed
+- Fix some transcoding edge cases with NVIDIA and pixel format `yuv420p10le`
 
 ## [0.1.1-alpha] - 2021-10-10
 ### Added